refactor(icon-build-helpers): update test signatures with metadata
