+++ packages/icon-build-helpers/src/metadata/extensions/__tests__/deprecated-test.js
@@ -50,7 +50,10 @@ describe('deprecated', () => {
     await expect(
       Metadata.check({
         adapter: yml,
-        input: '/',
+        input: {
+          svg: '/svg',
+          extensions: '/',
+        },
         extensions: [extension],
       })
     ).rejects.toThrowErrorMatchingInlineSnapshot(
