+++ packages/icon-build-helpers/src/metadata/extensions/__tests__/icons-test.js
@@ -41,7 +41,10 @@ describe('icons', () => {
     await expect(
       Metadata.check({
         adapter: yml,
-        input: '/',
+        input: {
+          svg: '/svg',
+          extensions: '/',
+        },
         extensions: [extension],
       })
     ).rejects.toThrowErrorMatchingInlineSnapshot(`
@@ -74,7 +77,10 @@ describe('icons', () => {
     await expect(
       Metadata.check({
         adapter: yml,
-        input: '/',
+        input: {
+          svg: '/svg',
+          extensions: '/',
+        },
         extensions: [extension],
       })
     ).rejects.toThrowErrorMatchingInlineSnapshot(
@@ -100,7 +106,10 @@ describe('icons', () => {
     await expect(
       Metadata.check({
         adapter: yml,
-        input: '/',
+        input: {
+          svg: '/svg',
+          extensions: '/',
+        },
         extensions: [extension],
       })
     ).rejects.toThrowErrorMatchingInlineSnapshot(`
@@ -126,7 +135,10 @@ describe('icons', () => {
     await expect(
       Metadata.check({
         adapter: yml,
-        input: '/',
+        input: {
+          svg: '/svg',
+          extensions: '/',
+        },
         extensions: [extension],
       })
     ).rejects.toThrowErrorMatchingInlineSnapshot(
