+++ packages/icon-build-helpers/src/metadata/extensions/output/optimizer.js
@@ -186,10 +186,10 @@ const plugins = [
     removeRasterImages: false,
   },
   {
-    mergePaths: false,
+    mergePaths: true,
   },
   {
-    convertShapeToPath: false,
+    convertShapeToPath: true,
   },
   {
     sortAttrs: true,
